Broke up Game to make any Scanner input into its own class
